done: exception handling excluding authentication

+++ src/main/java/dlanaras/com/github/services/BookingService.java
@@ -7,8 +7,10 @@ import java.util.List;
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 import javax.persistence.EntityManager;
+import javax.persistence.TransactionRequiredException;
 import javax.transaction.Transactional;
 
+import dlanaras.com.github.exceptions.NullValueException;
 import dlanaras.com.github.models.Booking;
 import dlanaras.com.github.models.User;
 
@@ -35,17 +37,30 @@ public class BookingService {
     }
 
     @Transactional
-    public Booking updateBooking(Booking booking) {
-        return entityManager.merge(booking);
+    public Booking updateBooking(Booking booking) throws IllegalArgumentException, TransactionRequiredException {
+        try {
+            return entityManager.merge(booking);
+        } catch (IllegalArgumentException e) {
+            throw e;
+        } catch (TransactionRequiredException e) {
+            throw e;
+        }
     }
 
-    public void deleteBooking(Long id) {
+    public void deleteBooking(Long id) throws NullValueException {
         var entity = entityManager.find(Booking.class, id);
+        if (entity == null) {
+            throw new NullValueException("No user with id: " + id + " was found");
+        }
         entityManager.remove(entity);
     }
 
-    public Float getLatestBookingPrice(Long userId) {
+    public Float getLatestBookingPrice(Long userId) throws NullValueException {
         List<Booking> userBookings = this.findAll(userId);
+
+        if(userBookings == null) {
+            throw new NullValueException("No bookings for user: " + userId);
+        }
 
         Comparator<Booking> dateComparator = Comparator.comparing(Booking::getStartDate); // What the hell is this java, just use LINQ
         Arrays.sort((userBookings.stream().toArray(Booking[] ::new)), dateComparator);
